코드 수정 및 주석 추가

+++ src/main/java/kr/ac/skuniv/service/member/MemberService.java
@@ -15,7 +15,6 @@ import org.springframework.stereotype.Service;
 import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
 
 @Service
 public class MemberService {
@@ -32,16 +31,14 @@ public class MemberService {
 		this.passwordEncoder = passwordEncoder;
 	}
 	private String token = "";
-	//회원가입
-	public void signUp(MemberDto request, MemberRole roles) {
-			Member existMember = memberRepository.findById(request.getId());
 
-			//아아디 존재하면 에러 출력
-//			if(existMember != null) {
-//				throw new UserDefineException("이미 존재하는 아이디입니다.");
-//			}
-
-			Member member = request.toEntity();
+	/**
+	 * 회원가입
+	 * @param memberDto : 회원가입을 진행할 데이터
+	 * @param roles : 고객과 작가를 구분하기 위한 권한
+	 */
+	public void signUp(MemberDto memberDto, MemberRole roles) {
+			Member member = memberDto.toEntity();
 			member.setPassword(passwordEncoder.encode(member.getPassword()));
 
 			if(roles.equals(MemberRole.CLIENT)){
@@ -54,8 +51,12 @@ public class MemberService {
 			memberRepository.save(member);
 	}
 
-	//로그인
-	public String signIn(SignUpDto signUpDto, HttpServletResponse response) {
+	/**
+	 * 로그인
+	 * @param signUpDto : 로그인할 데이터 (ID, PASSWORD)
+	 * @param response : 로그인 정보를 cookie에 담기 위한 객체
+	 */
+	public void signIn(SignUpDto signUpDto, HttpServletResponse response) {
 
 		Member login = memberRepository.findById(signUpDto.getId());
 
@@ -72,25 +73,20 @@ public class MemberService {
 		Cookie cookie = new Cookie("user", jwtProvider.createToken(login.getId(), login.getRole()));
 		cookie.setMaxAge(60*60*24);
 		response.addCookie(cookie);
-
-		return "login success";
-		//return jwtProvider.createToken(login.getId(), login.getRole());
 	}
 
-	//내정보 수정
+	/**
+	 * 회원정보 수정
+	 * @param request : userId를 조회하기 위한 HttpServletRequest 객체
+	 * @param memberDto : 수정할 데이터
+	 */
 	public void updateMember(HttpServletRequest request, MemberDto memberDto) {
-		Cookie[] cookies = request.getCookies();
-		for(Cookie cookie : cookies){
-			if(cookie.getName().equals("user")){
-				token = cookie.getValue();
-			}
-		}
-		String userId = jwtProvider.getUserIdByToken(token); //아이디로 변환
+		String userId = getUserIdByToken(request);
 
 		Member member = memberRepository.findById(userId);
 
 		if(member == null)
-			throw new UserDefineException("존재하지 않는 회원입니다");
+			throw new UserDefineException("회원 정보를 수정할 수 없습니다.");
 
 		memberDto.setPassword(passwordEncoder.encode(memberDto.getPassword()));
 
@@ -100,31 +96,30 @@ public class MemberService {
 
 	}
 
-	//정보 삭제 -> 회원 탈퇴
+	/**
+	 * 회원 탈퇴
+	 * @param request : userId를 조회하기 위한 HttpServletRequest 객체
+	 */
 	public void deleteMember(HttpServletRequest request) {
-
-		Cookie[] cookies = request.getCookies();
-		for(Cookie cookie : cookies){
-			if(cookie.getName().equals("user")){
-				token = cookie.getValue();
-			}
-		}
-		String userId = jwtProvider.getUserIdByToken(token);
+		String userId = getUserIdByToken(request);
 
 		Member member = memberRepository.findById(userId);
-		
+
+		if(member == null)
+			throw new UserDefineException("회원 정보를 삭제할 수 없습니다.");
+
 		memberRepository.delete(member);
 		
 	}
 
+	/**
+	 * 회원 정보 조회
+	 * @param request : userId를 조회하기 위한 HttpServletRequest 객체
+	 * @return : 조회한 회원 정보
+	 */
+	public MemberDto getMemberInfo(HttpServletRequest request) {
+		String userId = getUserIdByToken(request);
 
-/*
-
-	//회원 정보 열람
-	public MemberDto memberInfo(HttpServletRequest request){
-		Cookie[] cookies = request.getCookies();
-		String token = cookies[0].getValue(); //user라는 쿠키의 값을 꺼내서
-		String userId = jwtProvider.getUserIdByToken(token);
 		Member member = memberRepository.findById(userId);
 
 		if(member == null)
@@ -139,48 +134,12 @@ public class MemberService {
 				.age(member.getAge())
 				.build();
 	}
-*/
 
-    public Boolean checkUserID(String userId) {
-		Member member = memberRepository.findById(userId);
-
-		//아아디 존재하면 에러 출력
-		if (member != null) {
-			logger.error("이미 존재하는 아이디입니다!!");
-			return false;
-		} else {
-			logger.info("아이디 중복 체크 완료");
-			return true;
-		}
-	}
-
-    public MemberDto getMemberInfo(HttpServletRequest request) {
-
-		Cookie[] cookies = request.getCookies();
-		for(Cookie cookie : cookies){
-			if(cookie.getName().equals("user")){
-				token = cookie.getValue();
-			}
-		}
-
-	    String userId = jwtProvider.getUserIdByToken(token); //아이디로 변환
-
-        Member member = memberRepository.findById(userId);
-
-        if(member == null)
-            throw new UserDefineException("존재하지 않는 회원입니다");
-
-        return MemberDto.builder()
-                .id(member.getId())
-                .name(member.getName())
-                .affiliation(member.getAffiliation())
-                .phone(member.getPhone())
-                .sex(member.getSex())
-                .age(member.getAge())
-                .build();
-    }
-
-	public void logout(HttpServletRequest request, HttpServletResponse response) throws IOException {
+	/**
+	 * 로그아웃
+	 * @param request : userId를 조회하기 위한 HttpServletRequest 객체
+	 */
+	public void logout(HttpServletRequest request){
 		Cookie[] cookies = request.getCookies();
 		for(Cookie cookie : cookies){
 			if(cookie.getName().equals("user")){
@@ -189,6 +148,39 @@ public class MemberService {
 			}
 		}
 		logger.info("Logout Successfully!!");
-		response.sendError(HttpServletResponse.SC_OK, "Logout Successfully");
+	}
+
+	/**
+	 * ID 중복 체크
+	 * @param userId : 중복 체크를 하고자 하는 userId
+	 * @return : 중복된 ID가 없으면 True
+	 */
+    public Boolean checkUserID(String userId) {
+		Member member = memberRepository.findById(userId);
+
+		//아아디 존재하면 에러 출력
+		if (member != null) {
+			throw new UserDefineException("이미 존재하는 아이디입니다!!");
+		} else {
+			logger.info("아이디 중복 체크 완료");
+		}
+		return true;
+	}
+
+
+	/**
+	 * token을 통해 userId를 가져오는 메소드
+	 * @param request : userId를 조회하기 위한 HttpServletRequest 객체
+	 * @return userId
+	 * @throws UserDefineException
+	 */
+	private String getUserIdByToken(HttpServletRequest request) throws UserDefineException {
+		Cookie[] cookies = request.getCookies();
+		for(Cookie cookie : cookies){
+			if(cookie.getName().equals("user")){
+				token = cookie.getValue();
+			}
+		}
+		return jwtProvider.getUserIdByToken(token);
 	}
 }
